fix: layout cleanup
