+++ middleware.ts
@@ -9,5 +9,5 @@ export function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: ['/((?!_next/static|robots.txt|sitemap.xml|favicon.png|image.png|api).*)', '/'],
+  matcher: ['/((?!_next/static|robots.txt|sitemap.xml|favicon.png|image.jpg|api).*)', '/'],
 };
